refactor: change file name of crypto_port to ss_crypto
