+++ lib/ss_crypto.h
